Drop support for JDK < 17.
Scala 3 artifacts cannot be built on lower JDKs. Dealing with two
JDK versions is too annoying and error-prone, in particular when
publishing.
