+++ partest-suite/src/test/resources/scala/tools/partest/scalajs/2.13.18/run/t6827.check
@@ -1,0 +1,15 @@
+start at -5: java.lang.ArrayIndexOutOfBoundsException: -5
+start at -1: java.lang.ArrayIndexOutOfBoundsException: -1
+start at limit: ok
+start at limit-1: ok
+first 10: ok
+read all: ok
+test huge len: ok
+5 from 5: ok
+20 from 5: ok
+test len overflow: ok
+start beyond limit: ok
+read 0: ok
+read -1: ok
+invalid read 0: ok
+invalid read -1: ok
