+++ test-suite/shared/src/test/scala/org/scalajs/testsuite/compiler/RegressionTest.scala
@@ -910,18 +910,13 @@ class RegressionTest {
 
   @Test
   def inferConstableOrConstantDesc_Issue4545(): Unit = {
-    /* Depending on the JDK version used to compile this test, the types
-     * inferred for the arrays will change. On JDK 12+, they will involve
-     * Constable and/or ConstantDesc.
-     */
-
-    // On JDK 12+, both Constable and ConstantDesc
+    // Both Constable and ConstantDesc
     val b = Array("foo", java.lang.Integer.valueOf(5))
     assertEquals(2, b.length)
     assertEquals("foo", b(0))
     assertEquals(5, b(1))
 
-    // On JDK 15+, both Constable but Boolean is not a ConstantDesc
+    // Both Constable but Boolean is not a ConstantDesc
     val a = Array("foo", java.lang.Boolean.TRUE)
     assertEquals(2, a.length)
     assertEquals("foo", a(0))
