+++ test-suite/shared/src/test/scala/org/scalajs/testsuite/javalib/lang/ConstableTest.scala
