+++ test-suite/shared/src/test/scala/org/scalajs/testsuite/javalib/lang/ConstantDescTest.scala
