+++ test-suite/shared/src/test/scala/org/scalajs/testsuite/javalib/util/random/RandomGeneratorTest.scala
