+++ test-suite/shared/src/test/scala/org/scalajs/testsuite/javalib/util/regex/RegexMatcherTest.scala
@@ -12,6 +12,7 @@
 
 package org.scalajs.testsuite.javalib.util.regex
 
+import java.lang.StringBuilder
 import java.util.regex._
 
 import org.junit.Test
@@ -559,5 +560,34 @@ class RegexMatcherTest {
     assertEquals("abc", ms.next())
     assertEquals("b", ms.group("Bee"))
     assertFalse(ms.hasNext)
+  }
+
+  @Test def appendReplacementAndAppendTailStringBuilder(): Unit = {
+    // From the JavaDoc
+    val matcher = Pattern.compile("cat").matcher("one cat two cats in the yard")
+    val sb = new StringBuilder()
+
+    while (matcher.find()) {
+      matcher.appendReplacement(sb, "dog")
+    }
+    matcher.appendTail(sb)
+
+    assertEquals("one dog two dogs in the yard", sb.toString)
+  }
+
+  @Test def replaceAllWithReplacer(): Unit = {
+    val matcher = Pattern
+      .compile("a*b")
+      .matcher("aabfooaabfooabfoob")
+      .region(0, 3) // should be discarded
+    assertEquals("0foo6foo12foo17", matcher.replaceAll(m => m.start().toString()))
+  }
+
+  @Test def replaceFirstWithReplacer(): Unit = {
+    val matcher = Pattern
+      .compile("dog")
+      .matcher("zzzdogzzzdogzzz")
+      .region(0, 3) // should be discarded
+    assertEquals("zzz3zzzdogzzz", matcher.replaceFirst(m => m.start().toString()))
   }
 }
